@@ -3,32 +3,45 @@
   "version": "0.1.0",
   "description": "",
   "private": true,
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "prepare": "husky && husky install",
-    "prestart": "npm run compile:templates && npm run compile:forms && npm run compile:header && npm run compile:fields",
-    "start": "DEBUG=* && node server/server.js",
-    "compile:templates": "handlebars src/components/Cards/cards.hbs -f src/components/Cards/cards.precompiled.js",
-    "compile:forms": "handlebars src/components/Forms/forms.hbs -f src/components/Forms/forms.precompiled.js",
-    "compile:header": "handlebars src/components/Header/header.hbs -f src/components/Header/header.precompiled.js",
-    "compile:fields": "npm run compile:allfields && npm run compile:textfield && npm run compile:listfield",
-    "compile:allfields": "handlebars src/components/Fields/fields.hbs -f src/components/Fields/fields.precompiled.js",
-    "compile:textfield": "handlebars src/components/Fields/TextField/textfield.hbs -f src/components/Fields/TextField/textfield.precompiled.js",
-    "compile:listfield": "handlebars src/components/Fields/ListField/listfield.hbs -f src/components/Fields/ListField/listfield.precompiled.js",
+    "start": "webpack serve",
     "lint": "eslint . --fix && prettier --write .",
     "build": "webpack"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "eslint": "^9.11.0",
+    "@babel/core": "^7.25.8",
+    "@babel/preset-env": "^7.25.8",
+    "@babel/preset-typescript": "^7.25.7",
+    "@eslint/js": "^9.12.0",
+    "@stylistic/eslint-plugin-js": "^2.9.0",
+    "@types/eslint__js": "^8.42.3",
+    "babel-loader": "^9.2.1",
+    "css-loader": "^7.1.2",
+    "eslint": "^9.12.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
+    "handlebars-loader": "^1.7.3",
+    "html-webpack-plugin": "^5.6.0",
     "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
+    "postcss": "^8.4.47",
+    "postcss-loader": "^8.1.1",
+    "postcss-preset-env": "^10.0.7",
     "prettier": "3.3.3",
+    "sass": "^1.79.4",
+    "sass-loader": "^16.0.2",
+    "style-loader": "^4.0.0",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.6.3",
+    "typescript-eslint": "^8.10.0",
     "webpack": "^5.95.0",
-    "webpack-cli": "^5.1.4"
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^5.1.0"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx,json,html,css,scss,md}": [
