@@ -1,6 +1,9 @@
 import Navigator from '../../router/navigator.js';
 import Ajax from '../../ajax/ajax.js';
 import UserApi from '../../api/api_user.js';
+import FormsTemplate from '../Forms/forms.hbs';
+import styles from './forms.css';
+import indexStyles from '../../index.css';
 
 export default class Forms {
   constructor(parent) {
@@ -31,8 +34,8 @@ export default class Forms {
   render() {
     const placeForHeader = document.querySelector('.place-for-header');
     placeForHeader.innerHTML = '';
-    const template = Handlebars.templates['forms.hbs'];
-    this.parent.innerHTML = template({ context: this.context });
+    // const template = Handlebars.templates['forms.hbs'];
+    this.parent.innerHTML = FormsTemplate({ context: this.context });
     const AuthForm = document.querySelector('.auth-form-inputs');
 
     AuthForm.addEventListener('submit', this.handleSubmit.bind(this));
